requirements.txt
bcrypt
mysql-connector-python
hg+https://bitbucket.org/lcdr/pyraknet@4251284791e1476f29e51bbda44986e38a448aed